requirements.txt
asgiref==3.5.2
Django==3.2.6
django-bootstrap-form==3.4
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
